Add wrapper types to move from raw integers to types.
Values that would previously return reference time-scaled times as u64
will now return a new type: Instant.  This is not the same Instant from
std::time, but provides many of the same methods, and will at least
provide methods to calculate deltas, etc, that spit out
std::time::Duration.  As well, the type has a method for extracting the
raw u64 value for manual manipulation.

This is a move towards being able to better differentiate between the
timebase of the source and the reference clocks, as well as provide more
natural interoperability with the std::time API.

Also cleaned up some dependencies, including avoiding an implicit
dependency on metrics-core, instead moving it to be a conditional.