clock: change `Clock::now` to take an `&self` receiver
Currently, `Clock::now` takes an `&mut self` receiver. This is because
the last `now` value stored in the clock must be updated when using the
`Counter` clock type. However, since the `now` value is a single `u64`,
it's possible to implement this using `AtomicU64::fetch_max` instead,
allowing `Clock::now` to take a shared reference rather than a mutable
reference.

This requires changing `ClockType` to use a manual `Clone` impl, as
`atomic_shim::AtomicU64` doesn't implement `Clone`.

This does have one potential downside, which is that `atomic_shim` may
fall back to locking on platforms _without_ native 64-bit atomics. If
that's a serious issue, we might prefer not to make this change.
However, this lets targets with 64-bit atomics avoid a layer of locking
that would otherwise *always* be necessary if `Clock::now` is used
without a mutable reference to the clock available.

Since this changes a public method signature, this should be considered
a breaking change.

Let me know what you think!

Signed-off-by: Eliza Weisman <eliza@buoyant.io>