fix: switch to clock_gettime(3) for macOS/iOS for monotonic clock
This fixes an issue where on M1 (and presumably M2/M3) hardware,
the usage of mach_absolute_time() resulted in undermeasuring the
passage of time when compared to std::time::Instant. Now we read
from the monotonic clock (using clock_gettime, but conceptually
the same as calling mach_continuous_time()) which properly tracks
the passage of time regardless of device sleep, frequency, etc.